add new Broker Widgets

+++ aspnet-core/src/TACHYON.Application.Shared/Dashboards/Broker/Dto/ActorsNumbersDto.cs
@@ -2,9 +2,9 @@ namespace TACHYON.Dashboards.Broker
 {
     public class ActorsNumbersDto
     {
-        public int CarrierActorPercentage { get; set; }
+        public double CarrierActorPercentage { get; set; }
         
-        public int ShipperActorPercentage { get; set; }
+        public double ShipperActorPercentage { get; set; }
 
         public int TotalActors { get; set; }
     }
